[FIX] Move login and register component under user
Login and register component are user related and therefore should place under the user folder.
