+++ src/main/java/Clips/backend/user/registration/RegistrationRequest.java
@@ -1,4 +1,4 @@
-package Clips.backend.registration;
+package Clips.backend.user.registration;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
